gedis: split file into smaller ones and linting
